fix: improve logging maintainability and temporarily remove json logging configuration  (#146)
Refactor logging code for improved maintainability by temporarily removing JSON logging functionality.

Update the .gitignore to exclude .dcm and .sqlite files.

Introduce an environment variable for debug logging and set the default log level based on this variable.

- **New Features**
- Introduced a new `LoggingManager` class for structured logging
management.
- Added support for JSON logging configuration (currently commented
out).
- Added a new environment variable for logging level configuration in
development.
	- New entries in `.gitignore` to ignore `.dcm` and `.sqlite` files.
- Enhanced CI/CD pipeline with new jobs for building and publishing
documentation.

- **Improvements**
- Enhanced logging control by utilizing environment variables for log
levels.
- Updated logging directory structure and default log filename for
better organization.
	- Improved management of logging levels based on verbosity settings.

- **Bug Fixes**
- Improved handling of logging level discrepancies with user-specified
settings.

- **Documentation**
- Updated docstrings and comments for clarity on new logging setup and
features.
