+++ .github/workflows/main.yml
@@ -43,7 +43,7 @@ jobs:
           path: coverage-report
 
   ################################################################################################
-  # Unit-Tests: Run unit tests using pytest
+  # Build-Docs: Build documentation using mkdocs
   ################################################################################################
   Build-Docs:
     runs-on: ${{ matrix.os }}
@@ -64,9 +64,38 @@ jobs:
           cache: true
           locked: false # wont be the same because of the tag
 
-      - name: Run pytest
+      - name: Test docs build
         run: |
           pixi run -e ${{ matrix.env }} doc-build
+
+  ################################################################################################
+  # Publish-Docs: Publish documentation to GitHub Pages using mike only on main branch
+  ################################################################################################
+  Publish-Docs:
+    needs: Build-Docs
+    if: github.ref == 'refs/heads/main' 
+    runs-on: ${{ matrix.os }}
+    strategy:
+      matrix:
+        os: [ubuntu-latest]
+        env: ["docs"]
+
+    steps:
+      - name: Checkout code
+        uses: actions/checkout@v4
+
+      - name: Install Pixi
+        uses: prefix-dev/setup-pixi@v0.8.1
+        with:
+          environments: ${{ matrix.env }}
+          pixi-version: v0.36.0
+          cache: true
+          locked: false # wont be the same because of the tag
+
+      - name: Publish docs using mike
+        run: |
+          pixi run -e ${{ matrix.env }} mike deploy --push --update-aliases dev latest
+
 
   ################################################################################################
   # Codecov: Run codecov to check coverage
