+++ src/imgtools/logging/__init__.py
@@ -1,320 +1,15 @@
-import json as jsonlib
-import logging.config
 import os
 from pathlib import Path
-from typing import TYPE_CHECKING, Dict, List
 
 import structlog
-from structlog.processors import CallsiteParameter, CallsiteParameterAdder
 
-from imgtools.logging.processors import (
-	CallPrettifier,
-	ESTTimeStamper,
-	PathPrettifier,
-)
+from imgtools.logging.logging_config import DEFAULT_LOG_LEVEL, LoggingManager
 
-if TYPE_CHECKING:
-	from structlog.typing import Processor
+# Set the default log level from the environment variable or use the default
+DEFAULT_OR_ENV = os.environ.get('IMGTOOLS_LOG_LEVEL', DEFAULT_LOG_LEVEL)
 
-DEFAULT_LOG_LEVEL = 'WARNING'
-LOG_DIR_NAME = '.logs'
-DEFAULT_LOG_FILENAME = 'application.log'
-VALID_LOG_LEVELS = {'DEBUG', 'INFO', 'WARNING', 'ERROR', 'CRITICAL'}
-
-
-class LoggingManager:
-	"""
-	Manages the configuration and initialization of a structured logger.
-
-	This class provides flexible options for configuring log levels, formats,
-	and output destinations. Logs can be human-readable or in JSON format for
-	automated systems.
-
-	Parameters
-	----------
-	name : str
-	    Name of the logger instance.
-	base_dir : Path, optional
-	    Base directory for relative path prettification. Defaults to the
-	    current working directory.
-
-	Attributes
-	----------
-	name : str
-	    Name of the logger instance.
-	base_dir : Path
-	    Base directory for path prettification.
-	level : str
-	    Log level for the logger.
-	json_logging : bool
-	    Whether JSON logging is enabled.
-
-	Methods
-	-------
-	get_logger()
-	    Retrieve the configured logger instance.
-	configure_logging(json_logging=None, level=None)
-	    Dynamically adjust logging settings.
-
-	Raises
-	------
-	ValueError
-	    If an invalid log level is provided.
-	RuntimeError
-	    If the log directory or file cannot be created.
-
-	Examples
-	--------
-	Initialize with default settings:
-	    >>> manager = LoggingManager(name='mypackage')
-	    >>> logger = manager.get_logger()
-	    >>> logger.info('Info message')
-
-	Enable JSON logging:
-	    >>> manager = LoggingManager(name='mypackage', json_logging=True)
-	    >>> logger = manager.get_logger()
-	"""
-
-	def __init__(
-		self,
-		name: str,
-		base_dir: Path | None = None,
-	) -> None:
-		self.name = name
-		self.base_dir = base_dir or Path.cwd()
-		self.level = self._get_env_variable(f'{name}_LOG_LEVEL', DEFAULT_LOG_LEVEL).upper()
-		self.json_logging = (
-			self._get_env_variable(f'{name}_JSON_LOGGING', 'false').lower() == 'true'
-		)
-		self._initialize_logger()
-
-	@staticmethod
-	def _get_env_variable(variable: str, default: str) -> str:
-		"""
-		Retrieve an environment variable value with a fallback default.
-
-		Parameters
-		----------
-		variable : str
-		    The name of the environment variable.
-		default : str
-		    The default value if the environment variable is not set.
-
-		Returns
-		-------
-		str
-		    The value of the environment variable or the default.
-		"""
-		return os.environ.get(variable.upper(), default)
-
-	def _setup_json_logging(self) -> str:
-		"""
-		Set up the logging configuration for JSON output.
-
-		Ensures that the log directory exists and the log file is writable.
-
-		Returns
-		-------
-		str
-		    The path to the JSON log file.
-
-		Raises
-		------
-		PermissionError
-		    If the log file is not writable.
-		RuntimeError
-		    If the log directory cannot be created.
-		"""
-		try:
-			log_dir = self.base_dir / LOG_DIR_NAME
-			log_dir.mkdir(parents=True, exist_ok=True)
-			json_log_file = log_dir / DEFAULT_LOG_FILENAME
-			if json_log_file.exists() and not os.access(json_log_file, os.W_OK):
-				msg = f'Log file {json_log_file} is not writable'
-				raise PermissionError(msg)
-		except (PermissionError, OSError) as err:
-			msg = f'Failed to create log directory at {log_dir}: {err}'
-			raise RuntimeError(msg) from err
-		return str(json_log_file)
-
-	def _create_base_logging_config(self, pre_chain: List) -> Dict:
-		"""
-		Create the basic logging configuration without JSON-specific settings.
-
-		Parameters
-		----------
-		pre_chain : list
-		    List of processors for structured logging.
-
-		Returns
-		-------
-		dict
-		    Base logging configuration.
-		"""
-		return {
-			'version': 1,
-			'disable_existing_loggers': False,
-			'formatters': {
-				'console': {
-					'()': structlog.stdlib.ProcessorFormatter,
-					'processors': [
-						CallPrettifier(concise=True),
-						structlog.stdlib.ProcessorFormatter.remove_processors_meta,
-						structlog.dev.ConsoleRenderer(
-							exception_formatter=structlog.dev.RichTracebackFormatter(
-								width=-1, show_locals=False
-							),
-						),
-					],
-					'foreign_pre_chain': pre_chain,
-				},
-			},
-			'handlers': {
-				'console': {
-					'class': 'logging.StreamHandler',
-					'formatter': 'console',
-				},
-			},
-			'loggers': {
-				self.name: {
-					'handlers': ['console'],
-					'level': self.level,
-					'propagate': False,
-				},
-			},
-		}
-
-	def _add_json_logging_config(
-		self, logging_config: Dict, pre_chain: List, json_log_file: str
-	) -> Dict:
-		"""
-		Add JSON logging settings to the logging configuration.
-
-		Parameters
-		----------
-		logging_config : dict
-		    Existing logging configuration.
-		pre_chain : list
-		    List of processors for structured logging.
-		json_log_file : str
-		    Path to the JSON log file.
-
-		Returns
-		-------
-		dict
-		    Updated logging configuration.
-		"""
-		json_formatter = {
-			'json': {
-				'()': structlog.stdlib.ProcessorFormatter,
-				'processors': [
-					CallPrettifier(concise=False),
-					structlog.stdlib.ProcessorFormatter.remove_processors_meta,
-					structlog.processors.dict_tracebacks,
-					structlog.processors.JSONRenderer(serializer=jsonlib.dumps, indent=2),
-				],
-				'foreign_pre_chain': pre_chain,
-			},
-		}
-		json_handler = {
-			'json': {
-				'class': 'logging.handlers.RotatingFileHandler',
-				'formatter': 'json',
-				'filename': json_log_file,
-				'maxBytes': 10485760,
-				'backupCount': 5,
-			},
-		}
-
-		logging_config['formatters'].update(json_formatter)
-		logging_config['handlers'].update(json_handler)
-		logging_config['loggers'][self.name]['handlers'].append('json')
-		return logging_config
-
-	def _initialize_logger(self) -> None:
-		"""
-		Initialize the logger with the current configuration.
-		"""
-		pre_chain: List[Processor] = [
-			structlog.stdlib.add_log_level,
-			ESTTimeStamper(),
-			structlog.stdlib.add_logger_name,
-			structlog.stdlib.PositionalArgumentsFormatter(),
-			CallsiteParameterAdder(
-				[
-					CallsiteParameter.MODULE,
-					CallsiteParameter.FUNC_NAME,
-					CallsiteParameter.LINENO,
-				]
-			),
-			PathPrettifier(base_dir=self.base_dir),
-			structlog.stdlib.ExtraAdder(),
-			structlog.processors.StackInfoRenderer(),
-		]
-
-		logging_config = self._create_base_logging_config(pre_chain)
-
-		if self.json_logging:
-			json_log_file = self._setup_json_logging()
-			logging_config = self._add_json_logging_config(logging_config, pre_chain, json_log_file)
-
-		logging.config.dictConfig(logging_config)
-		structlog.configure(
-			processors=[
-				*pre_chain,
-				structlog.stdlib.ProcessorFormatter.wrap_for_formatter,
-			],
-			logger_factory=structlog.stdlib.LoggerFactory(),
-			wrapper_class=structlog.stdlib.BoundLogger,
-			cache_logger_on_first_use=True,
-		)
-
-	def get_logger(self) -> structlog.stdlib.BoundLogger:
-		"""
-		Retrieve the logger instance.
-
-		Returns
-		-------
-		structlog.stdlib.BoundLogger
-		    Configured logger instance.
-		"""
-		return structlog.get_logger(self.name)
-
-	def configure_logging(
-		self, json_logging: bool = False, level: str = DEFAULT_LOG_LEVEL
-	) -> structlog.stdlib.BoundLogger:
-		"""
-		Dynamically adjust logging settings.
-
-		Parameters
-		----------
-		json_logging : bool, default=False
-		    Enable or disable JSON logging.
-		level : str, optional
-		    Set the log level.
-
-		Returns
-		-------
-		structlog.stdlib.BoundLogger
-		    Updated logger instance.
-
-		Raises
-		------
-		ValueError
-		    If an invalid log level is specified.
-		"""
-		if level is not None:
-			if level not in VALID_LOG_LEVELS:
-				msg = f'Invalid logging level: {level}'
-				raise ValueError(msg)
-			self.level = level.upper()
-
-		if json_logging is not None:
-			self.json_logging = json_logging
-		self._initialize_logger()
-
-		return self.get_logger()
-
+LOG_DIR_NAME = Path('.imgtools/logs')
+DEFAULT_LOG_FILENAME = 'imgtools.log'
 
 def get_logger(name: str, level: str = 'INFO') -> structlog.stdlib.BoundLogger:
 	"""
@@ -333,16 +28,16 @@ def get_logger(name: str, level: str = 'INFO') -> structlog.stdlib.BoundLogger:
 	    Configured logger instance.
 	"""
 	logging_manager = LoggingManager(name)
-	env_level = os.environ.get(f'{name}_LOG_LEVEL', None)
-	if env_level != level and env_level is not None:
+	env_level = logging_manager.env_level
+
+	if env_level != level and env_level != DEFAULT_LOG_LEVEL:
 		logging_manager.get_logger().warning(
 			f'Environment variable {name}_LOG_LEVEL is {env_level} '
 			f'but you are setting it to {level}'
 		)
 	return logging_manager.configure_logging(level=level)
 
-
-logger = get_logger('imgtools', DEFAULT_LOG_LEVEL)
+logger = get_logger('imgtools', DEFAULT_OR_ENV)
 
 
 if __name__ == '__main__':
